# Performance improvements for formatting analyzer