--- a/src/EditorFeatures/CSharp/Formatting/Indentation/CSharpIndentationService.cs
+++ b/src/EditorFeatures/CSharp/Formatting/Indentation/CSharpIndentationService.cs
@@ -25,15 +25,15 @@ namespace Microsoft.CodeAnalysis.Editor.CSharp.Formatting.Indentation
     [ExportLanguageService(typeof(ISynchronousIndentationService), LanguageNames.CSharp), Shared]
     internal partial class CSharpIndentationService : AbstractIndentationService<CompilationUnitSyntax>
     {
-        private static readonly IFormattingRule s_instance = new FormattingRule();
+        private static readonly AbstractFormattingRule s_instance = new FormattingRule();
 
-        protected override IFormattingRule GetSpecializedIndentationFormattingRule()
+        protected override AbstractFormattingRule GetSpecializedIndentationFormattingRule()
         {
             return s_instance;
         }
 
         protected override AbstractIndenter GetIndenter(
-            ISyntaxFactsService syntaxFacts, SyntaxTree syntaxTree, TextLine lineToBeIndented, IEnumerable<IFormattingRule> formattingRules, OptionSet optionSet, CancellationToken cancellationToken)
+            ISyntaxFactsService syntaxFacts, SyntaxTree syntaxTree, TextLine lineToBeIndented, IEnumerable<AbstractFormattingRule> formattingRules, OptionSet optionSet, CancellationToken cancellationToken)
         {
             return new Indenter(
                 syntaxFacts, syntaxTree, formattingRules,
@@ -41,7 +41,7 @@ namespace Microsoft.CodeAnalysis.Editor.CSharp.Formatting.Indentation
         }
 
         public static bool ShouldUseSmartTokenFormatterInsteadOfIndenter(
-            IEnumerable<IFormattingRule> formattingRules,
+            IEnumerable<AbstractFormattingRule> formattingRules,
             CompilationUnitSyntax root,
             TextLine line,
             OptionSet optionSet,