… src/EditorFeatures/CSharp/Formatting/CSharpEditorFormattingService.cs
… src/EditorFeatures/CSharp/Formatting/Indentation/SmartTokenFormatter.cs
… src/EditorFeatures/CSharp/Formatting/Indentation/SmartTokenFormatterCommandHandler.cs
… src/EditorFeatures/CSharp/Formatting/TypingFormattingRule.cs
… src/EditorFeatures/CSharpTest/Formatting/FormattingEngineTestBase.cs
… src/EditorFeatures/Core/Implementation/AutomaticCompletion/Extensions.cs
… src/EditorFeatures/Core/Implementation/Formatting/Indentation/AbstractSmartTokenFormatterCommandHandler.cs
… src/EditorFeatures/Core/Implementation/SmartIndent/AbstractIndentationService.AbstractIndenter.cs
… src/EditorFeatures/Core/Implementation/SmartIndent/AbstractIndentationService.cs
… src/EditorFeatures/Core/Shared/Extensions/ITextSnapshotExtensions.cs
… src/EditorFeatures/TestUtilities/TestExportProvider.cs
… src/EditorFeatures/TestUtilities/Workspaces/TestFormattingRuleFactoryServiceFactory.cs
… src/EditorFeatures/VisualBasic/Formatting/Indentation/SmartTokenFormatter.vb
… src/EditorFeatures/VisualBasic/Formatting/Indentation/SmartTokenFormatterCommandHandler.vb
… src/EditorFeatures/VisualBasic/Formatting/Indentation/SpecialFormattingOperation.vb
… src/EditorFeatures/VisualBasic/Formatting/Indentation/VisualBasicIndentationService.Indenter.vb
… src/EditorFeatures/VisualBasic/Formatting/Indentation/VisualBasicIndentationService.vb
… src/EditorFeatures/VisualBasic/LineCommit/CommitFormatter.vb
… src/EditorFeatures/VisualBasic/NavigationBar/AbstractGenerateCodeItem.vb
… src/EditorFeatures/VisualBasic/Utilities/LineAdjustmentFormattingRule.vb
… src/EditorFeatures/VisualBasic/Wrapping/SeparatedSyntaxList/AbstractVisualBasicSeparatedSyntaxListWrapper.vb
… src/Features/CSharp/Portable/ChangeSignature/CSharpChangeSignatureService.cs
… src/Features/CSharp/Portable/ChangeSignature/ChangeSignatureFormattingRule.cs
… src/Features/CSharp/Portable/ExtractMethod/CSharpMethodExtractor.FormattingProvider.cs
… src/Features/CSharp/Portable/ExtractMethod/CSharpMethodExtractor.cs
… src/Features/CSharp/Portable/MetadataAsSource/CSharpMetadataAsSourceService.cs
… src/Features/CSharp/Portable/UseAutoProperty/CSharpUseAutoPropertyCodeFixProvider.cs
… src/Features/CSharp/Portable/UseConditionalExpression/CSharpUseConditionalExpressionForAssignmentCodeRefactoringProvider.cs
… src/Features/CSharp/Portable/UseConditionalExpression/CSharpUseConditionalExpressionForReturnCodeRefactoringProvider.cs
… src/Features/CSharp/Portable/UseConditionalExpression/MultiLineConditionalExpressionFormattingRule.cs