Recommended review strategy: review each commit.

On request I can easily break up the pull request to simplify review of individual changes, but keep in mind nearly all of the *observable* performance improvement comes from the final changes in the PR.

Times prior to this change:

* 46150ms
* 46215ms
* 47086ms

Times with `IFormattingRule` replaced with `AbstractFormattingRule` (eliminate interface dispatch):

* 46215ms
* 45208ms
* 45355ms

<details><summary>Times with virtual dispatch and elimination of <tt>ThreadLocal&lt;List&lt;T&gt;&gt;</tt>:</summary>

Moved to #32952 

* 45618ms
* 44991ms
* 46964ms

</details>

<details><summary>Times with above plus <tt>IndentationOptions</tt>:</summary>

Change is not included in this pull request.

* 48296ms
* 45684ms
* 44475ms

</details>

<br/>

Times with generics expanded to per-operation non-generic types.

* 40041ms
* 40309ms
* 41543ms