@sharwell 2019-01-27
> Details on the level of perf improvement?

No details yet. Need to set up test scenarios and get measurements. These changes are based on profiling a build of Roslyn.sln.