@sharwell 2019-01-30
> Details on the level of perf improvement?

Added to top.

This change reduced the CPU time *in the formatting code* from about 308 seconds to about 260 seconds.